\docType{package}
\name{svgg}
\alias{svgg}
\alias{svgg-package}
\title{svgg.}
\format{\preformatted{proto object 
 $ draw        :function (., data, scales, coordinates, na.rm = FALSE, ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 11 11 20 3 11 3 11 20 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0af709e4>  
 $ default_stat:function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 36 19 36 42 19 42 36 36 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0af709e4>  
 $ objname     : chr "point" 
 $ default_aes :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 38 18 44 3 18 3 38 44 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0af709e4>  
 $ draw_groups :function (., ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 10 18 10 45 18 45 10 10 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0af709e4>  
 $ required_aes: chr [1:2] "x" "y" 
 $ draw_legend :function (., data, ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 22 18 34 3 18 3 22 34 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0af709e4>  
 parent: proto object 
 .. parent: proto object 
}}
\usage{
GeomPoint
}
\description{
svgg.
}
\keyword{datasets}

