\docType{package}
\name{svgg}
\alias{svgg}
\alias{svgg-package}
\title{svgg.}
\format{\preformatted{proto object 
 $ draw        :function (., data, scales, coordinates, na.rm = FALSE, ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 11 11 22 3 11 3 11 22 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0b0579ec>  
 $ default_stat:function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 38 19 38 42 19 42 38 38 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0b0579ec>  
 $ objname     : chr "point" 
 $ default_aes :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 40 18 46 3 18 3 40 46 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0b0579ec>  
 $ draw_groups :function (., ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 10 18 10 45 18 45 10 10 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0b0579ec>  
 $ required_aes: chr [1:2] "x" "y" 
 $ draw_legend :function (., data, ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 24 18 36 3 18 3 24 36 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0b0579ec>  
 parent: proto object 
 .. parent: proto object 
}}
\usage{
GeomPoint
}
\description{
svgg.
}
\keyword{datasets}

